App runs but still needs to be debugged cause the desired functionality is missing.
	new file:   app/src/main/java/de/tudarmstadt/tk/umundoim/SubscriptionListActivity.java
The activity that provides a list of static possible trends to sub to.

	modified:   app/src/main/java/de/tudarmstadt/tk/umundoim/activity/ChatMainActivity.java
Here now we handle the returning result from subscriptions and add them to a spinner in which a user can chose which trend to post to.

	modified:   app/src/main/java/de/tudarmstadt/tk/umundoim/constant/Constants.java
I think this file is pretty much useless now. The scope of this project is far smaller than I imagined at first.(but we can keep it for the time being)